video: framebuffer properties must be 32bit values

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@783 f158a5a8-5612-0410-a976-696ce0be7e32

+++ drivers/vga_vbe.c
@@ -161,7 +161,7 @@ void vga_vbe_init(const char *path, unsigned long fb, uint32_t fb_size,
 	set_int_property(ph, "height", height);
 	set_int_property(ph, "depth", depth);
 	set_int_property(ph, "linebytes", linebytes);
-	set_int_property(ph, "address", fb & ~0x0000000F);
+	set_int_property(ph, "address", (u32)(fb & ~0x0000000F));
 
 	chosen = find_dev("/chosen");
 	push_str(path);
